requirements.txt
asgiref==3.2.7
biopython==1.76
Django==3.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
numpy==1.18.2
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
